Big feature: Implemented `NonNull` for `AllocatedMemory`.
Because of that, I also tried to get rid of most pointers and use `as_mut` and `as_ref` of `NonNull` instead. Because the pointers are known to be not null, we can dereference them automatically. This makes our life much easier during development and also makes the code more readable.