Changing and setting the permission of pages works now.
Fixed a bug: When handling the NPT fault, we **must not** advance the instruction pointer. We are essentially skipping to the next instruction from the current instruction that hasn't even been executed because the page wasn't available.

Other features implemented in this commit:
- Updated `AccessType` to be a normal enum
- Changing page permissions
- Mapping pages upon page fault

There's still lots of code optimizations to do to remove the duplicated code when setting up the NPT.