Bug fix: Removed custom data because it was causing crashes.
I'm not quite sure why. The crash originated in `processor_data.rs` when assigning the custom_data to the instance. I have still no idea why it was causing the crash but I figured it would be easier to just hardcode the rdtsc specific fields and don't allow custom data.